requirements.txt
httpx
psycopg2-binary 
openai>=1.0.0
